requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.5,<6.0
psycopg[binary]>=3.2,<4.0
django-environ>=0.11,<1.0
django-cors-headers>=4.6,<5.0
celery>=5.4,<6.0
redis>=5.2,<7.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
argon2-cffi>=23.1,<26.0
django-axes>=7.0,<8.0
django-ratelimit>=4.1,<5.0
drf-spectacular>=0.28,<1.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0

pyotp>=2.9,<3.0
stripe>=12.0,<13.0
